1. 3DS gui is now exactly like DS gui, just scaled. 2. 3DS Lid and backlights function like DS 3. Fixed 3DS Find IP button. 4. Cleaned up main9.cpp. 5. Implemented 3DS UDP non-blocking. 6. Increased stylus data from uint8_t to uint16_t for the enlarged 3DS screen. 7. Removed outdated PPJoy and replaced with current VJoy.

+++ client/arm9/source/gui/object.cpp
@@ -59,7 +59,7 @@ bool Object::IsVisible()
 {
 	return m_visible;
 }
-bool Object::IsClicked(uint8_t x, uint8_t y)
+bool Object::IsClicked(uint16_t x, uint16_t y)
 {
 	return m_rect.PointIntersect(x, y);
 }
